pyproject.toml
[project]
name = "sparse-generalization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Aryan Garg", email = "aryangarg794@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "black>=26.3.1",
    "dill>=0.4.0",
    "gymnasium>=1.2.2",
    "hydra-core>=1.3.2",
    "ipykernel>=7.1.0",
    "lightning>=2.5.6",
    "matplotlib>=3.10.7",
    "minigrid>=3.0.0",
    "networkx>=3.5",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "plotly>=6.5.0",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
    "torchsummary>=1.5.1",
    "torchvision>=0.24.1",
    "typing>=3.10.0.0",
    "wandb>=0.23.0",
]

[project.scripts]
sparse-generalization = "sparse_generalization:main"
train-toy = "sparse_generalization.scripts.train_toy:main"
box-world = "sparse_generalization.scripts.train_boxworld:main"
shapes = "sparse_generalization.scripts.train_shapes:main"
oracle = "sparse_generalization.scripts.run_oracle:main"
spartan = "sparse_generalization.scripts.run_spartan:main"
generate-box = "sparse_generalization.scripts.generate_boxworld_dataset:generate_dataset"
generate-box-v2 = "sparse_generalization.scripts.gen_box_v2:generate_dataset"
generate-shapes = "sparse_generalization.scripts.shape_dataset:generate_dataset"

[build-system]
requires = ["uv_build>=0.9.3,<0.10.0"]
build-backend = "uv_build"

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
